BROKEN COMMIT Added simple "comment" parsing to catch strings like:
#This is a comment.

This currently forces the application to deal with comments, so I'm contacting the json_spirit developer to try to fix it. Also, there will undoubtedly be some problems with #'s inside normal json strings. But I'd rather iron those out later; comments are very necessary now, and I need them to work.

This commit is broken because the application logic does not exist to handle comments. I want to see what the original author has to say first.