Added window functionality:   - Width/height/client width/height, default width/height.   - Added expand/resize/move window functions.
Now at the stage where the drawing routines need to be added. This will be slightly messy, but I suppose it's ok...

Expecting some bugs on the first few runs, as "SetWindowPos", "MoveWindow", and a few other variations of the two probably won't interract  nicely (and were just happily and tediously balanced in earlier versions of WZ).