Finally got windows and hotkeys to be set as necessary.
This calls the setXX() functions multiple times, even when not necessary, but it limits calls to ONLY when the input system has crossed a boundary (sentence went from non-empty to empty, roman string went from non-empty to empty, both cases vice versa, help mode active/deactivate).

And the code's a lot shorter this way, and easy to understand. Now, to check if the Roman/Letter inputs actually maintain their internal state properly... then to finally compile this monster of a change.