Superficial improvement: reduced candidate's constructor list.
Actually, matching rules is fine. However, the way we handle "matched strings" and dots for variables is silly. We should take advantage of the fact that every letter of the input string WILL be matched somewhere, rather than constantly rebuilding it. Also, the recursive nature of apply() is messing things up.