Added a demo Zawgyi keyboard. However, sequences like "ua" are switching into their Uncode forms; ZWS doesn't seem to be working right.
We'll need a full review of the layout, which shouldn't be too hard, and we'll definitely need some testers who are familiar with typing Zawgyi. 

Also, I'd guess that the "Zawgyi Key" (~) doesn't work, since we take "no string change" to mean "nothing matched". We'll need to fix this (Ayar needs it too for English typing).