BROKEN COMMIT: Started adding the close/minimize buttons to the Onscreen Keyboard.
This will require a bit of re-working (e.g., don't try to draw buttons when minimzied, remember to keep closed if closed) but shouldn't be too bad.

We'll also need an "OnHighlight()" function ---it doesn't need to track OnMouseOut or anything fancy like that; perhaps OnMouseOver is enough. (We could do OnMouseOut, or we could have some kind of global "NoMouseFound" event).